refactor(dh): move emptyValue pipe to shared folder
