+++ libs/dh/metering-point/feature-overview/src/lib/primary-master-data/dh-metering-point-primary-master-data.component.ts
@@ -34,7 +34,7 @@ import {
   WattIconSize,
 } from '@energinet-datahub/watt';
 
-import { DhEmptyValuePipeScam } from './empty-value.pipe';
+import { DhEmptyValuePipeScam } from '../shared/empty-value.pipe';
 import { emDash } from '../shared/em-dash';
 import { DhYesNoPipeScam } from '../shared/yes-no.pipe';
